backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-docx==1.1.2
openpyxl==3.1.5
ruff==0.6.9
black==24.8.0
isort==5.13.2